:sparkles:feat:Make only USerRole.Owner create store
