+++ src/main/java/sipozizo/tabling/common/config/SecurityConfig.java
@@ -34,6 +34,7 @@ public class SecurityConfig {
                 .addFilterBefore(jwtFilter, SecurityContextHolderAwareRequestFilter.class)
                 .authorizeHttpRequests(auth -> auth
                         .requestMatchers("/api/auth/login", "/api/auth/register").permitAll()
+                        .requestMatchers("/api/stores/owners").hasRole("OWNER")
                         .anyRequest().authenticated())
                 .build();
     }
